Update typescript to allow build.

+++ app/api/like/route.tsx
@@ -2,35 +2,46 @@ import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
 import { authOptions } from "../auth/[...nextauth]/route";
 import prisma from "@/prisma/client";
+import session from "@/app/interface/session";
 
 export async function POST(request: NextRequest) {
   const body = await request.json()
 
-  const session = await getServerSession(authOptions)
+  const session: session | null = await getServerSession(authOptions)
 
-  const dbCall = await prisma.likes.create({
-    data: {
-      tweetId: body.id,
-      likerId: session.session.user.id
-    }
-  })
+  if (session) {
+    const dbCall = await prisma.likes.create({
+      data: {
+        tweetId: body.id,
+        likerId: session.session.user.id
+      }
+    })
+  
+    return NextResponse.json(dbCall)
+  }
 
-  return NextResponse.json({"done": "done"})
+  return NextResponse.json({})
+
 }
 
 export async function DELETE(request: NextRequest) {
   const body = await request.json()
 
-  const session = await getServerSession(authOptions)
+  const session: session | null = await getServerSession(authOptions)
 
-  const dbCall = await prisma.likes.delete({
-    where: {
-      likerId_tweetId: {
-        tweetId: body.id,
-        likerId: session.session.user.id
+  if (session) {
+    const dbCall = await prisma.likes.delete({
+      where: {
+        likerId_tweetId: {
+          tweetId: body.id,
+          likerId: session.session.user.id
+        }
       }
-    }
-  })
+    })
+  
+    return NextResponse.json(dbCall)
+  }
 
-  return NextResponse.json({"done": "done"})
+  return NextResponse.json({})
+
 }
